--- src/main/java/roomescape/reservation/controller/ReservationPageController.java
@@ -1,19 +1,0 @@
-package roomescape.reservation.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class ReservationPageController {
-
-    @GetMapping("/reservation")
-    public String userReservationPage() {
-        return "reservation";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String adminReservationPage() {
-        return "admin/reservation-new";
-    }
-}
-
